buying and selling stocks now changes cash balance

+++ app/src/main/java/com/keith/stocksim/AddOrderFragment.java
@@ -32,8 +32,21 @@ import retrofit2.converter.gson.GsonConverterFactory;
 
 public class AddOrderFragment extends Fragment {
     DatabaseHandler db;
+    void displayToast(final String msg) {
+        new Handler(Looper.getMainLooper()).post(new Runnable() {
+            @Override
+            public void run() {
+                Context context = getContext();
+                int duration = Toast.LENGTH_SHORT;
+
+                Toast toast = Toast.makeText(context, msg, duration);
+                toast.show();
+            }
+        });
+    }
     public void addOrder(){
         db =  ((MainActivity) getActivity()).db;
+        final SharedPreferences sharedPreferences = ((MainActivity) getActivity()).sharedPreferences;
         EditText tickerEdit =(EditText) getView().findViewById(R.id.tickerEdit);
         final String tickerSymbol = tickerEdit.getText().toString().toUpperCase();
         if (tickerSymbol.equals("")) {
@@ -52,17 +65,32 @@ public class AddOrderFragment extends Fragment {
                 try {
                     Call<StockQuery> theQuote = service.getQuote(tickerSymbol);
                     Response<StockQuery> response = theQuote.execute();
+                    Integer oldShares = (db.getStock(tickerSymbol) != null) ? db.getStock(tickerSymbol).numShares : 0;
+                    // ticker is valid
                     if (response.body() != null) {
                         EditText sharesEdit = (EditText) getView().findViewById(R.id.orderSizeEdit);
+                        if (sharesEdit.getText().toString().equals("")) return;
                         Integer newShares = Integer.parseInt(sharesEdit.getText().toString());
-                        Integer oldShares = (db.getStock(tickerSymbol) != null)? db.getStock(tickerSymbol).numShares:0;
-                        Integer totalShares = newShares + oldShares;
-                        if (totalShares == 0) {
-                            db.deleteStock(new Stock(tickerSymbol, 0,0));
-                            return;
+                        if (newShares == 0) return;
+                        if (response.body().quote.latestPrice * newShares > sharedPreferences.getFloat("cashBalance", 0)) {
+                            displayToast("You do not have enough money");
                         }
-                        if (db.getStock(tickerSymbol) == null) db.addStock(new Stock(tickerSymbol, totalShares, 0));
-                        else db.updateStock(new Stock(tickerSymbol, totalShares, 0));
+                        else if(oldShares + newShares < 0){
+                            displayToast("You Do Not Have Enough Shares");
+                        }
+                        else{
+                            Integer totalShares = newShares + oldShares;
+                            if (totalShares == 0) {
+                                db.deleteStock(new Stock(tickerSymbol, 0, 0));
+                            }
+                            else if (db.getStock(tickerSymbol) == null)
+                                db.addStock(new Stock(tickerSymbol, totalShares, 0));
+                            else db.updateStock(new Stock(tickerSymbol, totalShares, 0));
+                            SharedPreferences.Editor editor = sharedPreferences.edit();
+                            editor.putFloat("cashBalance", sharedPreferences.getFloat("cashBalance", 0) - newShares * (float)response.body().quote.latestPrice);
+                            editor.commit();
+                            displayToast("Order Added Successfully");
+                        }
                     } else {
                         new Handler(Looper.getMainLooper()).post(new Runnable() {
                             @Override
